chore: package 구조 잘못된 부분 수정
