+++ core/core-api/src/main/kotlin/com/lms/api/problem/ProblemController.kt
@@ -1,9 +1,9 @@
 package com.lms.api.problem
 
-import com.lms.core.domain.problem.domain.problem.request.ProblemGetRequest
-import com.lms.core.domain.problem.domain.problem.response.ProblemFilterResponse
-import com.lms.core.domain.problem.service.ProblemGetService
-import com.lms.core.response.ApiResponse
+import com.lms.core_domain.problem.domain.request.ProblemGetRequest
+import com.lms.core_domain.problem.domain.response.ProblemFilterResponse
+import com.lms.core_domain.problem.service.ProblemGetService
+import com.lms.core_domain.response.ApiResponse
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.ModelAttribute
