--- core/core-common/src/main/kotlin/com/lms/core/enum/ProblemType.kt
@@ -1,11 +1,0 @@
-package com.lms.core.enum
-
-enum class ProblemType {
-    ALL,
-    SUBJECTIVE,
-    SELECTION;
-
-    fun isSame(other: ProblemType): Boolean {
-        return this == other
-    }
-}
