--- core/core-common/src/main/kotlin/com/lms/core/enum/exception/InvalidLevelValueException.kt
@@ -1,8 +1,0 @@
-package com.lms.core.enum.exception
-
-import com.lms.core.exception.BusinessException
-
-class InvalidLevelValueException(
-    levelValue: Int,
-    message: String = "Invalid level value $levelValue"
-) : BusinessException(message)
