--- core/core-common/src/main/kotlin/com/lms/core/exception/BusinessException.kt
@@ -1,8 +1,0 @@
-package com.lms.core.exception
-
-import org.springframework.http.HttpStatus
-
-open class BusinessException(
-    override val message: String,
-    val status: HttpStatus = HttpStatus.BAD_REQUEST
-) : RuntimeException(message)
