+++ core/core-common/src/main/kotlin/com/lms/core_domain/enum/ProblemType.kt
@@ -1,0 +1,11 @@
+package com.lms.core_domain.enum
+
+enum class ProblemType {
+    ALL,
+    SUBJECTIVE,
+    SELECTION;
+
+    fun isSame(other: ProblemType): Boolean {
+        return this == other
+    }
+}
