+++ core/core-common/src/main/kotlin/com/lms/core_domain/enum/exception/InvalidLevelValueException.kt
@@ -1,0 +1,8 @@
+package com.lms.core_domain.enum.exception
+
+import com.lms.core_domain.exception.BusinessException
+
+class InvalidLevelValueException(
+    levelValue: Int,
+    message: String = "Invalid level value $levelValue"
+) : BusinessException(message)
