+++ core/core-common/src/main/kotlin/com/lms/core_domain/exception/BusinessException.kt
@@ -1,0 +1,8 @@
+package com.lms.core_domain.exception
+
+import org.springframework.http.HttpStatus
+
+open class BusinessException(
+    override val message: String,
+    val status: HttpStatus = HttpStatus.BAD_REQUEST
+) : RuntimeException(message)
