+++ core/core-common/src/test/kotlin/com/lms/enum/ProblemTypeTest.kt
@@ -1,6 +1,6 @@
 package com.lms.enum
 
-import com.lms.core.enum.ProblemType
+import com.lms.core_domain.enum.ProblemType
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
