--- core/core-domain/src/main/kotlin/com/lms/core/domain/problem/domain/problem/request/ProblemGetRequest.kt
@@ -1,11 +1,0 @@
-package com.lms.core.domain.problem.domain.problem.request
-
-import com.lms.core.enum.Level
-import com.lms.core.enum.ProblemType
-
-data class ProblemGetRequest(
-    val totalCount: Int,
-    val unitCodeList: List<String>,
-    val level: Level,
-    val problemType: ProblemType
-)
