--- core/core-domain/src/main/kotlin/com/lms/core/domain/problem/domain/problem/response/ProblemFilterResponse.kt
@@ -1,11 +1,0 @@
-package com.lms.core.domain.problem.domain.problem.response
-
-import com.lms.core.enum.ProblemType
-
-data class ProblemFilterResponse(
-    val id: Long,
-    val answer: String,
-    val unitCode: String,
-    val level: Int,
-    val problemType: ProblemType,
-)
