+++ core/core-domain/src/main/kotlin/com/lms/core_domain/problem/domain/exception/InvalidValueException.kt
@@ -1,0 +1,7 @@
+package com.lms.core_domain.problem.domain.exception
+
+import com.lms.core_domain.exception.BusinessException
+
+class InvalidValueException(
+    message: String = "Invalid value exception",
+) : BusinessException(message)
