+++ storage/db-core/src/main/kotlin/com/lms/db/core/problem/ProblemEntity.kt
@@ -1,7 +1,7 @@
 package com.lms.db.core.problem
 
-import com.lms.core.domain.problem.domain.problem.Problem
-import com.lms.core.enum.ProblemType
+import com.lms.core_domain.enum.ProblemType
+import com.lms.core_domain.problem.domain.Problem
 import com.lms.db.core.config.BaseEntity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
